renaming module transforms to generators
