+++ badgers/generators/geolocated_data/__init__.py
