+++ badgers/generators/image/__init__.py
