+++ badgers/generators/time_series/__init__.py
