+++ badgers/generators/time_series/noise.py
