+++ badgers/generators/time_series/outliers.py
