+++ tests/transforms/tabular_data/test_missingness.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from numpy.random import default_rng
 
-from badgers.transforms.tabular_data.missingness import MissingValueGenerator
+from badgers.generators.tabular_data.missingness import MissingValueGenerator
 from tests.transforms.tabular_data import generate_test_data_only_features
