+++ tests/transforms/tabular_data/test_outliers.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import numpy as np
 from numpy.random import default_rng
 
-from badgers.transforms.tabular_data.outliers import OutliersTransformer, ZScoreSampling, HistogramSampling
+from badgers.generators.tabular_data.outliers import OutliersTransformer, ZScoreSampling, HistogramSampling
 from tests.transforms.tabular_data import generate_test_data_only_features
